+++ operator-common/src/main/java/io/strimzi/operator/common/TimeoutException.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.operator.common.operator.resource;
+package io.strimzi.operator.common;
 
 /**
  * Thrown to indicate that timeout has been exceeded.
