+++ operator-common/src/main/java/io/strimzi/operator/common/operator/resource/concurrent/ResourceSupport.java
@@ -16,8 +16,8 @@ import io.fabric8.kubernetes.client.dsl.Listable;
 import io.fabric8.kubernetes.client.dsl.Watchable;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.ReconciliationLogger;
+import io.strimzi.operator.common.TimeoutException;
 import io.strimzi.operator.common.Util;
-import io.strimzi.operator.common.operator.resource.TimeoutException;
 
 import java.io.Closeable;
 import java.io.IOException;
